chore: move utils outside e2e
This commit moves the utils folder from tests/e2e to tests.
This also reverts the source code go.mod file back to the original one
before cert-manager was added.
