+++ codecov.yml
@@ -1,3 +1,3 @@
 ignore:
   - pkg/utils/k8s/           # Exclude all utils related to k8s
-  - tests/e2e/               # Exclude the e2e test directory
+  - tests/               # Exclude the e2e test directory
