+++ tests/e2e/oc_literal_test.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"github.com/sustainable.computing.io/kepler-operator/tests/e2e/utils/oc"
+	"github.com/sustainable.computing.io/kepler-operator/tests/utils/oc"
 )
 
 const podSpecTemplate = `
