+++ tests/e2e/power_monitor_internal_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/sustainable.computing.io/kepler-operator/internal/controller"
 	powermonitor "github.com/sustainable.computing.io/kepler-operator/pkg/components/power-monitor"
 	"github.com/sustainable.computing.io/kepler-operator/pkg/utils/k8s"
-	"github.com/sustainable.computing.io/kepler-operator/tests/e2e/utils"
+	"github.com/sustainable.computing.io/kepler-operator/tests/utils"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 )
