+++ tests/e2e/power_monitor_test.go
@@ -19,7 +19,7 @@ import (
 
 	powermonitor "github.com/sustainable.computing.io/kepler-operator/pkg/components/power-monitor"
 	"github.com/sustainable.computing.io/kepler-operator/pkg/utils/k8s"
-	"github.com/sustainable.computing.io/kepler-operator/tests/e2e/utils"
+	"github.com/sustainable.computing.io/kepler-operator/tests/utils"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
