+++ tests/utils/oc/literal.go
