+++ tests/utils/oc/literal_test.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"github.com/sustainable.computing.io/kepler-operator/tests/e2e/utils/oc"
+	"github.com/sustainable.computing.io/kepler-operator/tests/utils/oc"
 )
 
 func Test_Literal(t *testing.T) {
