+++ tests/utils/power_monitor_internal_builder.go
